requirements/local.txt
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
email-validator==1.1.1
fastapi==0.65.2
passlib==1.7.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
mysqlclient==1.4.4
requests==2.24.0
SQLAlchemy==1.3.18
uvicorn==0.14.0
uvloop==0.14.0, >=0.14.0
websockets==9.1
Faker==0.8.10
mysql-connector==2.1.7
mysqlclient==1.4.4